finish assignments

+++ src/rong/algorithm/Q18_4Sum.java
@@ -32,7 +32,7 @@ public class Q18_4Sum {
                 int k = j + 1;
                 int l = nums.length-1;
                 while (k < l){
-                    int sum = nums[i] +nums[j] +nums[k] +nums[l];
+                    long sum = (long)nums[i] +nums[j] +nums[k] +nums[l];
                     if (sum == target){
                         result.add(Arrays.asList(nums[i],nums[j],nums[k],nums[l]));
                         k++;
